Fix python indenting error on line 6405 of CRStools/test.ipynb
In commit 8a822375 of 4MOST_CRS_tools in CRStools/

  jupyter nbconvert test.ipynb --to python

fails because of an indentation error. This is because line 6405 of
CRStools/test.ipynb with '@classmethod' has only two spaces instead of
four spaces.

This commit inserts two extra spaces to fix the indenting;
'jupyter nbconvert test.ipynb --to python' works for me after the
fix. This commit fixes bug #1 [1].

[1] https://github.com/4most-crs/4MOST_CRS_tools/issues/1